user authorization not working
